+++ R/ggpredict.R
@@ -45,8 +45,14 @@ utils::globalVariables(c("observed", "predicted"))
 #'          predictors. Residuals and observed values are set to \code{NA}.
 #'          Usually, this argument is only used internally by \code{ggaverage()}.
 #' @param typical Character vector, naming the function to be applied to the
-#'           covariates over which the effect is "averaged". The default is "mean".
-#'           See \code{\link[sjstats]{typical_value}} for options.
+#'          covariates over which the effect is "averaged". The default is "mean".
+#'          See \code{\link[sjstats]{typical_value}} for options.
+#' @param ppd Logical, if \code{TRUE}, predictions for \code{stanreg}-models are
+#'          based on the posterior predictive distribution
+#'          (\code{\link[rstanarm]{posterior_predict}}). If \code{FALSE} (the
+#'          default), predictions are based on posterior draws of the linear
+#'          predictor (\code{\link[rstanarm]{posterior_linpred}}). See 'Details'
+#'          for non-gaussian models used with \code{posterior_predict()}.
 #' @param ... Further arguments passed down to \code{predict()}.
 #'
 #' @details Currently supported model-objects are: \code{lm, glm, glm.nb, lme, lmer,
@@ -80,15 +86,24 @@ utils::globalVariables(c("observed", "predicted"))
 #'          marginal effects.
 #'          \cr \cr
 #'          \code{ggpredict()} also works with \strong{stanreg}-models from
-#'          the \CRANpkg{rstanarm}-package. The predictions are based on
-#'          \code{\link[rstanarm]{posterior_linpred}} and hence have some
-#'          limitations; the uncertainty of the error term is not taken into
+#'          the \CRANpkg{rstanarm}-package. The predicted values are the median
+#'          value of all drawn posterior samples. The confidence intervals for
+#'          \code{stanreg}-models are actually high density intervals, computed
+#'          by \code{\link[sjstats]{hdi}}. By default, the predictions are
+#'          based on \code{\link[rstanarm]{posterior_linpred}} and hence have some
+#'          limitations: the uncertainty of the error term is not taken into
 #'          account. The recommendation is to use the posterior predictive
 #'          distribution (\code{\link[rstanarm]{posterior_predict}}), however,
 #'          \code{posterior_linpred()} is faster and easier to compute (especially
-#'          for models with binary outcome). The confidence intervals for
-#'          \code{stanreg}-models are actually HDI, computed by
-#'          \code{\link[sjstats]{hdi}}.
+#'          for models with binary outcome). Use the argument \code{ppd = TRUE}
+#'          to compute predictions based on draws from the posterior predictive
+#'          distribution. Note that for binomial models, the \code{newdata}-argument
+#'          used in \code{posterior_predict()} must also contain the vector
+#'          with the number of trials. In this case, a dummy-vector is used,
+#'          where all values for the response are set to 1. Also, for non-gaussian
+#'          models and \code{ppd = TRUE}, no confidence intervals are calculated,
+#'          since this would require drawing many replicates of the posterior
+#'          predictive distributions.
 #'
 #' @note Since data for \code{ggaverage()} comes from the model frame, not all
 #'       possible combinations of values in \code{terms} might be present in the data,
@@ -221,7 +236,7 @@ utils::globalVariables(c("observed", "predicted"))
 #' @importFrom tibble has_name as_tibble
 #' @importFrom purrr map
 #' @export
-ggpredict <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), full.data = FALSE, typical = "mean", ...) {
+ggpredict <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), full.data = FALSE, typical = "mean", ppd = FALSE, ...) {
   # check arguments
   type <- match.arg(type)
 
@@ -234,15 +249,15 @@ ggpredict <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), full.dat
   }
 
   if (inherits(model, "list"))
-    purrr::map(model, ~ggpredict_helper(.x, terms, ci.lvl, type, full.data, typical, ...))
+    purrr::map(model, ~ggpredict_helper(.x, terms, ci.lvl, type, full.data, typical, ppd, ...))
   else
-    ggpredict_helper(model, terms, ci.lvl, type, full.data, typical, ...)
+    ggpredict_helper(model, terms, ci.lvl, type, full.data, typical, ppd, ...)
 }
 
 
 # workhorse that computes the predictions
 # and creates the tidy data frames
-ggpredict_helper <- function(model, terms, ci.lvl, type, full.data, typical, ...) {
+ggpredict_helper <- function(model, terms, ci.lvl, type, full.data, typical, ppd, ...) {
   # check class of fitted model
   fun <- get_predict_function(model)
 
@@ -276,7 +291,7 @@ ggpredict_helper <- function(model, terms, ci.lvl, type, full.data, typical, ...
 
 
   # compute predictions here -----
-  fitfram <- select_prediction_method(fun, model, expanded_frame, ci.lvl, type, binom_fam, ...)
+  fitfram <- select_prediction_method(fun, model, expanded_frame, ci.lvl, type, faminfo, ppd, ...)
 
 
   # init legend labels
@@ -399,6 +414,6 @@ ggpredict_helper <- function(model, terms, ci.lvl, type, full.data, typical, ...
 
 #' @rdname ggpredict
 #' @export
-mem <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), full.data = FALSE, typical = "mean", ...) {
-  ggpredict(model, terms, ci.lvl, type, full.data, typical, ...)
+mem <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), full.data = FALSE, typical = "mean", ppd = FALSE, ...) {
+  ggpredict(model, terms, ci.lvl, type, full.data, typical, ppd, ...)
 }